feat: schedule-account api

See merge request s12-fintech-finance-sub1/S12P21B110!77

+++ server/src/main/java/com/b110/jjeonchongmu/domain/account/dto/TransferRequestDTO.java
@@ -16,5 +16,5 @@ public class TransferRequestDTO {
 	private AccountType toAccountType;
 	private Long toAccountId;
 	private String tradeDetail;
-	private Integer transferAmount;
+	private Long transferAmount;
 }
